requirements.txt
django==1.8.5
django-daterange-filter==1.2.0
django-taggit==0.17.1
django-datetime-widget==0.9.3
